added motor rotation to BeverageMakerTask

+++ assignment-2/arduino/src/iot/actuator/display-i2c/DisplayI2C.cpp
@@ -1,7 +1,6 @@
 #include "DisplayI2C.h"
-#include <Arduino.h>
 
-DisplayI2C::DisplayI2C(int address, int width, int height) {
+DisplayI2C::DisplayI2C(uint8_t address, uint8_t width, uint8_t height) {
     this->address = address;
     this->lcd = LiquidCrystal_I2C(this->address, width, height);
 }
@@ -11,7 +10,7 @@ void DisplayI2C::setup() {
     lcd.backlight();
 }
 
-void DisplayI2C::changeCursor(int column, int row) {
+void DisplayI2C::changeCursor(uint8_t column, uint8_t row) {
     this->column = column;
     this->row = row;
 }
