+++ assignment-2/arduino/src/iot/actuator/motor/Motor.h
@@ -3,18 +3,19 @@
 #define _SERVO_H_
 #include "iot/actuator/actuator.h"
 #include <Servo.h>
+#include <Arduino.h>
 
 class Motor : public Actuator {
 
 private:
-    int pin;
-    int current_angle = 0;
+    uint8_t pin;
+    uint8_t current_angle = 0;
     Servo servo;
 
 public:
-    Motor(int pin);
+    Motor(uint8_t pin);
     void setup();
-    void rotateTo(int angle);
+    void rotateTo(uint8_t angle);
     void returnToStart();
     void goToEnd();
 };
